+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - pass the route parameter to flutter launch configurations
 - add the ability to launch mojo apps from the flutter launch configuration
 - added a getting started and plugin status view
+- add extract local refactoring
 
 ## 0.4.15
 - create hyperlinks from exception traces in the console view